feat: implement core functionality for insurance data retrieval and processing in Spain
- Added modules for fetching insurance companies and distributors from the Spanish regulator's website.
- Implemented data extraction utilities for parsing HTML and JavaScript data.
- Created models for insurance companies and customer service information.
- Introduced logging for better traceability and debugging.
- Established folder synchronization for storing insurance company data.
- Defined versioning and constants for API headers and URLs.
