+++ insurance_product_data_spain/constants/public_urls.py
